Add a testcase to test the gang management
The gang management now has a smaller granularity and a better management
of memory context, meanwhile, gangs are not pre-assigned anymore for main
plan and init plans. An issue was reported for old implementation of gang
management, although it cannot been reproduced now, it's still necessary
to add a test case to avoid regression in the feature.

The test strategy is preparing a query contains both main and init plan,
main plan create a unnamed portal and will need multiple slices , init
plan also need to allocated multiple slices in a named portal (eg: a
cursor), the expect result is gang management works fine in such mixed
combination.