Do a dependencies check when coverting a table to replicated table
system columns of replicated table should not be exposed to users,
we do the check at parser stage, however, if users create views or
rules involve the system columns of a hash distributed table and
then the hash distributed table is converted to a replicated table,
then users can still access the system columns of replicated table
through views and rules because they bypass the check in parser
stage. To resolve this, we add a dependencies check when altering
a table to replicated table, users need to drop the views or rules
first.

I tried to add a recheck in later stages like planner or executor,
but it seems that system columns like CTID is used internally for
basic DMLs like UPDATE/DELETE and those columns are added even
before the rewrite of views and rules, so adding a recheck will
block basic DMLs too.