Fix calculation of WorkfileMgrLock and WorkfileQuerySpaceLock
All lwlocks are stored in MainLWLockArray which is an array of
LWLockPadded structures:

typedef union LWLockPadded
{
  LWLock lock;
  char pad[LWLOCK_PADDED_SIZE];
} LWLockPadded;

The calculation in SyncHTPartLockId to fetch a lwlock is
incorrect because it offsets the array as an LWLock array.
In current code base, it works fine because the size of
LWLock happens to be 32, if structure LWLock get enlarged,
the calculation will mess up.