ExecSquelchNode and ExecEagerFreeXXX code structure modify
The main change of this commit is function structure refactor that
suggested by Heikki, it seems fragile that we have ExecSquelchNode()
and ExecEagerFreeXXX() calls added to all the places where we do
"return NULL". If a new place where these functions return is added
in the upstream, we will miss making these GPDB-specific calls, so
we added a few wrapper functions to match upstream more closely and
it does make code cleaner and easier to read.