Little refine of SendEosUDPIFC() to resolve potential issue
Previously, Even a connection has been explicitly set to inactive, the old code
might still treat the connection as active if conn->cdbProc is not null and
conn->sndQueue is not empty and increase activeCount. In the next loop, because
conn->stillActive is true, conn->unackQueue or conn->sndQueue are never be
freed and activeCount always be non-zeror and might cause an infinite loop.