initTM() should not change the role to superuser temporarily
When a GPDB server start, initTM() is called to goes through all
in-doubt transactions and resolve them, however, it need to be under
a superuser context somehow. Previously, GPDB uses a hacky way which
temporary change the role to superuser and then restore it back after
resolving in-doubt transactions, the old solution might involve security
issues in one hand, in another hand it strongly dependes on MyProcPort
structure which is not always available,

We know initTM() is always called when a server start after a crash or
PANIC and now the first 'QD' will try to call initTM(), so to fix this,
we let the first 'QD' whose authenticated user is superuser to do the
initTM(), we know startup process is always started and it is under
superuser context, so initTM() can be called for sure.