Fix an issue caused by mismatched targetlist of Motion node
A Motion node may has a subplan whose targetlist need to be expanded
to make sure all entries of the subplan's hashExpr are in its
targetlist, in this case, the motion node should also update its
targetlist, otherwise, the difference of targetlist will cause
Motion node parse the tuple (especially for MemTuple) incorrectly
with a mismatched tuple description, make wrong result or even crash.

It's doubtful of the intention to add subplan's flow hashExpr to the
targetlist of the subplan and removing it can also fix the issue, but
removing it seems to take more risk.

Co-authored-by: Adam Lee <ali@pivotal.io>