Backport the fix for invalidation of plan cache
In commit a453004ea194 to align with PG 8.3.23, GPDB backported commit ee33b95
from PG 8.4 which contain the modification of plan cache invalidation.
Unfortunately, the commit backported is buggy because it does not skip the
utility statements, so another commit e245c91 is merged to PG 8.4 to fix it,
however, that commit is not included in GPDB5 now.

Backported upstream commit:

commit e245c91211adb26be7cfdaeed2977411a003622f
Author: Tom Lane <tgl@sss.pgh.pa.us>
Date:   Wed Jan 13 16:57:03 2010 +0000

    Fix bug #5269: ResetPlanCache mustn't invalidate cached utility statements,
    especially not ROLLBACK.  ROLLBACK might need to be executed in an already
    aborted transaction, when there is no safe way to revalidate the plan.  But
    in general there's no point in marking utility statements invalid, since
    they have no plans in the normal sense of the word; so we might as well
    work a bit harder here to avoid future revalidation cycles.

    Back-patch to 8.4, where the bug was introduced.