Add parallel worker and hash flags to locus
Now a path can has paralle_workers > 0, we need to
re-consider the meanning of locus for partiationed
locus. For example, table foo is distributed by c1,
so table foo's locus is CdbLocusType_Hashed, the
distkey is c1, if we start two workers to scan the
table concurrently, from the workers's view, the
data will be randomly distributed, but for the
segment's view, the data is still distributed by
c1. This is usefull when you plan a join on a
partial_path and a normal path, if we know the
joinrestrictinfo matched the c1, we don't need
to add a motion for the join.

We add a new hash flag to a locus:
HASHED_ON_SEGMENT, HASHED_ON_WORKERS

For a baserel, if parallel_workers is zero, the
path locus has HASHED_ON_SEGMENT | HASHED_ON_WORKERS
hash flag, if parallel_workers is non-zero, the
path locus has only HASHED_ON_SEGMENT.

If a path is the result of redistibute motion,
the path will only have HASHED_ON_WORKERS flag.

This will be usefull to decide whether we need
to add a motion for JOIN or AGG.

Path also has parallel_workers, we still add
a new parallel_workers to CdbPathLocus because
GPDB always use CdbPathLocus to consider data
redistribution, so add parallel_workers to
CdbPathLocus will make things easier.