Enable parallelModeOK in GPDB planner
parallelModeOK is always set to false in GPDB, now we set it
to true if gp_enable_parallelscan is true, partial_pathlist
can contain the paths whose parallel_workers is non-zero.

There are some difference on partial_pathlist between GPDB
and postgresql.

GPDB have Motions to communicate between workers, which makes
GPDB more powerfull. In postgres, all data flow in workers
must be gathered to leader worker, which cause partial_pathlist
in postgres in not actually a valid path for the final rel, it
must be merged to pathlist with a GATHER node. In GPDB,
partial_pathlist also contains valid paths, it can be used by
final_rel by adding a Gather Motion, it is same with pathlist
in GPDB.

So in GPDB, we need to consider partial_pathlist in more
cases, but for now, we just merge partial_pathlist and pathlist
at a certain point. Added a FIXME to fix this.