Enhance gpcheckcat persistent check
gpcheckcat persistent check is consist of a series of extra/missing
check, one is to detect the extra relation files in the file system
after a table is dropped, the main idea is:
1. using gp_persistent_relation_node_check() to list all relation
   files in the filesystem
2. say whether catalog pg_class contains the entries referring to
   those relfilenodes.
3. say whether catalog pg_persistent_relation_node contains the
   entries referring to those relfilenodes.

Another background is:
heap tables only ever store segment_file_num 0 in the persistent
tables, while ao/co tables will store every segment_file_num that
they use, so we need to handle the difference in the SQL file,
otherwise, segment_file_num > 0 will be treated as extra files.

With the old filter in the SQL file, extra segment files > 0 cannot
be detected, so enhance the filter.