Fix flaky uao_crash_compaction_column test
In a normal vacuum, for each column, the vacuum will
1) create a new segment file and record it in the gp_aocsseg_xxxxx
   table with a new entry with frozen xmin.
2) compact the tuples in the old segment file and move it to the new
   segment file.
3) drop/truncate the old segment file and mark it as dropped in the
   gp_aocsseg_xxxxx table.

In the uao_crash_compaction_column test, a PANIC is injected between
step 1 and step2 on seg0, so the whole transaction is aborted.

For seg1 and seg2, because step1 is using a frozen xmin, the new segment
file is visible even the vacuum is aborted.

For seg0, things are different. Because the PANIC occurs right behind
the new segment file is created in a very short time, so the buffer cache
for the new record in the gp_aocsseg_xxxx and the buffer cache for the
related xlog have a big chance to not be flushed to the disk, so when the
seg0 is recovered, we will not see the record in the gp_aocsseg_xxxx. But
if the buffer is flushed to the disk because the checkpointer/bgwriter
before the panic, we will see the new segment file record in the
gp_aocsseg_xxxx table.

This is RCA of the flakiness.

Considering the purpose of the test case, this commit separates the query
to two queries:

1. "SELECT segno,column_num,physical_segno,tupcount,modcount,state FROM
   gp_toolkit.__gp_aocsseg('crash_vacuum_in_appendonly_insert') where
   segno = 1" to verify the old segment is still visible after the vacuum
   is aborted.
2. "SELECT sum(tupcount) FROM gp_toolkit.__gp_aocsseg('crash_vacuum_in_ap
   pendonly_insert') where state = 1 and segno = 2" to verify the new
   segment file contains no tuples.